# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API quản lý chuỗi trường mầm non. Lỗi trả về theo RFC 7807 (application/problem+json).",
    "title" : "Preschool Management API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/files/{id}/download-url" : {
      "get" : {
        "operationId" : "downloadUrl",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DownloadUrlResponse"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "summary" : "Xin link tải file (có hạn)",
        "tags" : [ "File" ]
      }
    },
...
    "/api/v1/staff/summary" : {
      "get" : {
        "operationId" : "summary",
        "parameters" : [ {
          "in" : "query",
          "name" : "schoolId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffSummary"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "summary" : "Tóm tắt nhân sự: tổng, theo vị trí, giấy tờ sắp hết hạn",
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{id}" : {
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffDetail"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "summary" : "Hồ sơ nhân viên (kèm quyền của người xem)",
        "tags" : [ "Nhân sự" ]
      },
      "put" : {
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StaffFields"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffDetail"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "summary" : "Sửa hồ sơ (không gồm cơ sở, trạng thái, lương, ngân hàng)",
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/certificates" : {
      "get" : {
        "operationId" : "certificates",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CertificateDto"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      },
      "post" : {
        "operationId" : "createCertificate",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CertificateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CertificateDto"
                }
              }
            },
            "description" : "Created"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/certificates/{id}" : {
      "delete" : {
        "operationId" : "deleteCertificate",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      },
      "put" : {
        "operationId" : "updateCertificate",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CertificateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CertificateDto"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/contracts" : {
      "get" : {
        "operationId" : "contracts",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ContractDto"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      },
      "post" : {
        "operationId" : "createContract",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContractRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContractDto"
                }
              }
            },
            "description" : "Created"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/contracts/{contractId}" : {
      "delete" : {
        "operationId" : "deleteContract",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "contractId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      },
      "put" : {
        "operationId" : "updateContract",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "contractId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContractRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContractDto"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/dependents" : {
      "get" : {
        "operationId" : "dependents",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DependentDto"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      },
      "post" : {
        "operationId" : "createDependent",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DependentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DependentDto"
                }
              }
            },
            "description" : "Created"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/dependents/{id}" : {
      "delete" : {
        "operationId" : "deleteDependent",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      },
      "put" : {
        "operationId" : "updateDependent",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DependentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DependentDto"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/documents" : {
      "get" : {
        "operationId" : "documents",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StaffDocumentDto"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "summary" : "Giấy tờ của nhân viên (mọi phiên bản, bản hiện hành có current = true)",
        "tags" : [ "Nhân sự" ]
      },
      "post" : {
        "operationId" : "addDocument",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StaffDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffDocumentDto"
                }
              }
            },
            "description" : "Created"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "summary" : "Tải lên giấy tờ (phiên bản mới, không ghi đè bản cũ)",
        "tags" : [ "Nhân sự" ]
      }
    },
    "/api/v1/staff/{staffId}/documents/{documentId}" : {
      "delete" : {
        "operationId" : "deleteDocument",
        "parameters" : [ {
          "in" : "path",
          "name" : "staffId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "documentId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "Cơ sở đang chọn (UUID). Bỏ trống = tất cả cơ sở trong phạm vi của người dùng.",
          "in" : "header",
          "name" : "X-School-Id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Lỗi (RFC 7807, thông điệp tiếng Việt)"
          }
        },
        "tags" : [ "Nhân sự" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DownloadUrlResponse" : {
        "type" : "object",
        "properties" : {
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "url" : {
            "type" : "string"
          }
        },
        "required" : [ "expiresAt", "url" ]
      },
...
      "Problem" : {
        "type" : "object",
        "description" : "Lỗi theo RFC 7807",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Mã lỗi máy đọc, ví dụ INVALID_CREDENTIALS"
          },
          "detail" : {
            "type" : "string",
            "description" : "Thông điệp tiếng Việt hiển thị cho người dùng"
          },
          "errors" : {
            "type" : "array",
            "description" : "Lỗi từng trường khi dữ liệu không hợp lệ",
            "items" : {
              "type" : "object",
              "properties" : {
                "field" : {
                  "type" : "string"
                },
                "message" : {
                  "type" : "string"
                }
              },
              "required" : [ "field", "message" ]
            }
          },
          "instance" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string",
            "description" : "Tiêu đề tiếng Việt theo mã HTTP"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "title", "status" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}